week02/day09/requirements.txt
gradio>=4.29.0
faiss-cpu>=1.8.0
chromadb>=0.5.4
sentence-transformers>=3.0.1
huggingface-hub>=0.24.0
numpy>=1.26.0
